+++ src/interpreters/Elixir_original.rs
@@ -1,19 +1,16 @@
 #[derive(Clone)]
 #[allow(non_camel_case_types)]
-pub struct Julia_original {
+pub struct Elixir_original {
     support_level: SupportLevel,
     data: DataHolder,
     code: String,
     main_file_path: String,
     cache_dir: String,
 
-    interpreter: String,
-    interpreter_args: Vec<String>, // for now, used for --project=....
-
     current_output_id: u32,
 }
 
-impl Julia_original {
+impl Elixir_original {
     fn wait_out_file(&self, path: String, id: u32) -> Result<String, String> {
         let end_mark = String::from("sniprun_finished_id=") + &id.to_string();
         let start_mark = String::from("sniprun_started_id=") + &id.to_string();
@@ -46,67 +43,55 @@ impl Julia_original {
             info!("not found yet");
         }
 
+        // elixir-specific filtering
+        let mut cleaned_contents = String::new();
+        let mut current_line = contents.lines().next().unwrap_or("");
+        for l in contents.lines() {
+            if current_line.starts_with("iex(") {
+                let index = current_line.rfind(")>").unwrap();
+                cleaned_contents += &current_line[index + 2..];
+            } else if !(current_line == ":ok" && l.starts_with("iex(")) {
+                // keep the line
+                cleaned_contents += current_line;
+                cleaned_contents += "\n";
+            } else {
+                // do nothing, (discarding the line)
+            }
+            current_line = l;
+        }
+        // info!("cleaned contents: {:?}", cleaned_contents);
+
+        let contents = cleaned_contents;
         let index = contents.rfind(&start_mark).unwrap();
         Ok(contents[index + start_mark.len()..contents.len() - end_mark.len() - 1].to_owned())
     }
-
-    fn fetch_config(&mut self) {
-        let default_interpreter = String::from("julia");
-        self.interpreter = default_interpreter;
-        if let Some(used_interpreter) =
-            Julia_original::get_interpreter_option(&self.get_data(), "interpreter")
-        {
-            if let Some(interpreter_string) = used_interpreter.as_str() {
-                info!("Using custom interpreter: {}", interpreter_string);
-                self.interpreter = interpreter_string.to_string();
-            }
-        }
-
-        if let Some(project_opt) =
-            Julia_original::get_interpreter_option(&self.get_data(), "project")
-        {
-            if let Some(mut project_str) = project_opt.as_str() {
-                if project_str == "." {
-                    info!(
-                        "Using neovim's current working directory as julia --project: {}",
-                        self.data.projectroot
-                    );
-                    project_str = &self.data.projectroot;
-                }
-
-                self.interpreter_args = vec![String::from("--project=") + project_str];
-            }
-        }
-    }
 }
 
-impl Interpreter for Julia_original {
-    fn new_with_level(data: DataHolder, level: SupportLevel) -> Box<Julia_original> {
+impl Interpreter for Elixir_original {
+    fn new_with_level(data: DataHolder, level: SupportLevel) -> Box<Elixir_original> {
         //create a subfolder in the cache folder
-        let rwd = data.work_dir.clone() + "/julia_original";
+        let rwd = data.work_dir.clone() + "/elixir_original";
         let mut builder = DirBuilder::new();
         builder.recursive(true);
         builder
             .create(&rwd)
-            .expect("Could not create directory for julia-original");
+            .expect("Could not create directory for elixir-original");
 
         //pre-create string pointing to main file's and binary's path
-        let mfp = rwd.clone() + "/main.jl";
+        let mfp = rwd.clone() + "/main.exs";
 
-        Box::new(Julia_original {
+        Box::new(Elixir_original {
             data,
             support_level: level,
             code: String::from(""),
             main_file_path: mfp,
-            interpreter: String::new(),
-            interpreter_args: Vec::new(),
             cache_dir: rwd,
             current_output_id: 0,
         })
     }
 
     fn get_name() -> String {
-        String::from("Julia_original")
+        String::from("Elixir_original")
     }
 
     fn default_for_filetype() -> bool {
@@ -118,14 +103,14 @@ impl Interpreter for Julia_original {
     }
 
     fn has_repl_capability() -> bool {
-        true
+        false
     }
 
     fn get_supported_languages() -> Vec<String> {
         vec![
-            String::from("Julia"),
-            String::from("julia"),
-            String::from("jl"),
+            String::from("Elixir"),
+            String::from("elixir"),
+            String::from("exs"),
         ]
     }
 
@@ -145,13 +130,12 @@ impl Interpreter for Julia_original {
     }
 
     fn check_cli_args(&self) -> Result<(), SniprunError> {
-        // All cli arguments are sendable to julia
+        // All cli arguments are sendable to python
         // Though they will be ignored in REPL mode
         Ok(())
     }
 
     fn fetch_code(&mut self) -> Result<(), SniprunError> {
-        self.fetch_config();
         if !self
             .data
             .current_bloc
@@ -176,12 +160,11 @@ impl Interpreter for Julia_original {
     fn build(&mut self) -> Result<(), SniprunError> {
         // info!("python code:\n {}", self.code);
         write(&self.main_file_path, &self.code)
-            .expect("Unable to write to file for julia_original");
+            .expect("Unable to write to file for elixir_original");
         Ok(())
     }
     fn execute(&mut self) -> Result<String, SniprunError> {
-        let output = Command::new(&self.interpreter)
-            .args(&self.interpreter_args)
+        let output = Command::new("elixir")
             .arg(&self.main_file_path)
             .args(&self.get_data().cli_args)
             .output()
@@ -189,7 +172,7 @@ impl Interpreter for Julia_original {
         if output.status.success() {
             Ok(String::from_utf8(output.stdout).unwrap())
         } else {
-            if Julia_original::error_truncate(&self.get_data()) == ErrTruncate::Short {
+            if Elixir_original::error_truncate(&self.get_data()) == ErrTruncate::Short {
                 return Err(SniprunError::RuntimeError(
                     String::from_utf8(output.stderr.clone())
                         .unwrap()
@@ -207,13 +190,13 @@ impl Interpreter for Julia_original {
     }
 }
 
-impl ReplLikeInterpreter for Julia_original {
+impl ReplLikeInterpreter for Elixir_original {
     fn fetch_code_repl(&mut self) -> Result<(), SniprunError> {
         self.fetch_code()?;
 
         if !self.read_previous_code().is_empty() {
             // nothing to do, kernel already running
-            info!("Julia kernel already running");
+            info!("elixir kernel already running");
 
             if let Some(id) = self.get_pid() {
                 // there is a race condition here but honestly you'd have to
@@ -229,8 +212,7 @@ impl ReplLikeInterpreter for Julia_original {
             // launch everything
             self.set_pid(0);
 
-            let init_repl_cmd = self.data.sniprun_root_dir.clone()
-                + "/src/interpreters/Julia_original/init_repl.sh";
+            let init_repl_cmd = self.data.sniprun_root_dir.clone() + "/ressources/init_repl.sh";
             info!(
                 "launching kernel : {:?} on {:?}",
                 init_repl_cmd, &self.cache_dir
@@ -242,18 +224,17 @@ impl ReplLikeInterpreter for Julia_original {
                         .args(&[
                             init_repl_cmd,
                             self.cache_dir.clone(),
-                            Julia_original::get_nvim_pid(&self.data),
-                            self.interpreter.clone(),
+                            Elixir_original::get_nvim_pid(&self.data),
+                            "iex".to_owned(),
                         ])
-                        .args(&self.interpreter_args)
                         .output()
                         .unwrap();
 
-                    return Err(SniprunError::CustomError("julia REPL exited".to_owned()));
+                    return Err(SniprunError::CustomError("elixir REPL exited".to_owned()));
                 }
                 Ok(Fork::Parent(_)) => {}
                 Err(_) => info!(
-                    "Julia_original could not fork itself to the background to launch the kernel"
+                    "Elixir_original could not fork itself to the background to launch the kernel"
                 ),
             };
 
@@ -262,27 +243,21 @@ impl ReplLikeInterpreter for Julia_original {
             self.save_code("kernel_launched".to_owned());
 
             Err(SniprunError::CustomError(
-                "Julia kernel launched, re-run your snippet".to_owned(),
+                "elixir kernel launched, re-run your snippet".to_owned(),
             ))
         }
     }
 
     fn add_boilerplate_repl(&mut self) -> Result<(), SniprunError> {
         self.add_boilerplate()?;
-        let start_mark = String::from("println(\"sniprun_started_id=")
+        let start_mark = String::from("IO.puts(\"sniprun_started_id=")
             + &self.current_output_id.to_string()
             + "\")\n";
-        let end_mark = String::from("println(\"sniprun_finished_id=")
+        let end_mark = String::from("IO.puts(\"sniprun_finished_id=")
             + &self.current_output_id.to_string()
             + "\")\n";
 
-        let start_mark_err = String::from("println(stderr, \"sniprun_started_id=")
-            + &self.current_output_id.to_string()
-            + "\")\n";
-        let end_mark_err = String::from("println(stderr, \"sniprun_finished_id=")
-            + &self.current_output_id.to_string()
-            + "\")\n";
-        self.code = start_mark + &start_mark_err + "\n" + &self.code + "\n" + &end_mark_err + &end_mark;
+        self.code = start_mark + &self.code + &String::from("\n") + &end_mark;
         Ok(())
     }
 
@@ -291,10 +266,10 @@ impl ReplLikeInterpreter for Julia_original {
     }
 
     fn execute_repl(&mut self) -> Result<String, SniprunError> {
-        info!("running launcher");
         let send_repl_cmd = self.data.sniprun_root_dir.clone() + "/ressources/launcher_repl.sh";
+        info!("running launcher (via {})", send_repl_cmd);
         let res = Command::new(send_repl_cmd)
-            .arg(self.main_file_path.clone())
+            .arg(self.cache_dir.clone() + "/main.exs")
             .arg(self.cache_dir.clone() + "/fifo_repl/pipe_in")
             .spawn()
             .expect("could not run launcher");
@@ -310,14 +285,14 @@ impl ReplLikeInterpreter for Julia_original {
 }
 
 #[cfg(test)]
-mod test_julia_original {
+mod test_elixir_original {
     use super::*;
 
     #[test]
     fn simple_print() {
         let mut data = DataHolder::new();
-        data.current_bloc = String::from("println(\"hello\")");
-        let mut interpreter = Julia_original::new(data);
+        data.current_bloc = String::from("IO.puts(\"hello\")");
+        let mut interpreter = Elixir_original::new(data);
         let res = interpreter.run();
 
         // should panic if not an Ok()